tools/orchestrator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cc-sdd-orchestrator"
version = "0.1.0"
description = "External orchestrator for cc-sdd implement/modify pipelines"
requires-python = ">=3.10"
dependencies = [
    "rich>=13.0",
    "anyio>=4.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["tools"]
